requirements.txt
baidu_aip==4.16.11
numpy==1.24.4
opencv_python==4.8.0.74
Requests==2.31.0
